Copilot: refactoring phase 1 ad 2
