+++ cmd/oceano-state-manager/fingerprint.go
@@ -1,179 +1,14 @@
 package main
 
-import (
-	"bytes"
-	"fmt"
-	"io"
-	"log"
-	"math/bits"
-	"os/exec"
-	"strconv"
-	"strings"
-)
+import internalrecognition "github.com/alemser/oceano-player/internal/recognition"
 
-// Fingerprint is the raw Chromaprint fingerprint as a slice of uint32 values,
-// one value per ~0.13 s of audio. Generated by fpcalc(1) with the -raw flag.
-type Fingerprint []uint32
+type Fingerprint = internalrecognition.Fingerprint
+type Fingerprinter = internalrecognition.Fingerprinter
 
-// ParseFingerprint parses the comma-separated uint32 output of fpcalc -raw.
-func ParseFingerprint(s string) (Fingerprint, error) {
-	s = strings.TrimSpace(s)
-	if s == "" {
-		return nil, fmt.Errorf("fingerprint: empty input")
-	}
-	fields := strings.Split(s, ",")
-	fp := make(Fingerprint, 0, len(fields))
-	for i, f := range fields {
-		v, err := strconv.ParseUint(strings.TrimSpace(f), 10, 32)
-		if err != nil {
-			return nil, fmt.Errorf("fingerprint: field %d: %w", i, err)
-		}
-		fp = append(fp, uint32(v))
-	}
-	return fp, nil
-}
+var ParseFingerprint = internalrecognition.ParseFingerprint
+var BER = internalrecognition.BER
+var GenerateFingerprints = internalrecognition.GenerateFingerprints
 
-// BER returns the minimum bit error rate between a and b over all time shifts
-// in [-maxShift, +maxShift]. Values below 0.35 indicate the same recording
-// (same threshold used by AcoustID). Returns 1.0 when either slice is empty.
-func BER(a, b Fingerprint, maxShift int) float64 {
-	if len(a) == 0 || len(b) == 0 {
-		return 1.0
-	}
-	best := 1.0
-	for shift := -maxShift; shift <= maxShift; shift++ {
-		if v := berAtShift(a, b, shift); v < best {
-			best = v
-		}
-	}
-	return best
-}
-
-// berAtShift computes BER with b shifted right by |shift| positions (or a
-// shifted right when shift is negative). Only the overlapping region is scored.
-func berAtShift(a, b Fingerprint, shift int) float64 {
-	aStart, bStart := 0, 0
-	if shift >= 0 {
-		bStart = shift
-	} else {
-		aStart = -shift
-	}
-	n := len(a) - aStart
-	if m := len(b) - bStart; m < n {
-		n = m
-	}
-	if n <= 0 {
-		return 1.0
-	}
-	var total int
-	for i := 0; i < n; i++ {
-		total += bits.OnesCount32(a[aStart+i] ^ b[bStart+i])
-	}
-	return float64(total) / float64(n*32)
-}
-
-// Fingerprinter generates a raw Chromaprint fingerprint from a WAV file.
-// Implementations must be safe for concurrent use.
-type Fingerprinter interface {
-	// Generate returns the raw fingerprint for the segment starting at
-	// offsetSec, lasting lengthSec seconds. A lengthSec that exceeds the
-	// file duration is safe — fpcalc processes whatever audio is available.
-	Generate(wavPath string, offsetSec, lengthSec int) (Fingerprint, error)
-}
-
-// fpcalcFingerprinter shells out to fpcalc(1) from the chromaprint-tools package.
-type fpcalcFingerprinter struct{}
-
-func (fpcalcFingerprinter) Generate(wavPath string, offsetSec, lengthSec int) (Fingerprint, error) {
-	if offsetSec == 0 {
-		out, err := exec.Command("fpcalc", "-raw", "-length", strconv.Itoa(lengthSec), wavPath).Output()
-		if err != nil {
-			return nil, fmt.Errorf("fpcalc: %w", err)
-		}
-		return parseFpcalcOutput(out)
-	}
-
-	// fpcalc does not support -offset; pipe ffmpeg-trimmed audio into fpcalc via stdin.
-	ffmpegCmd := exec.Command("ffmpeg", "-ss", strconv.Itoa(offsetSec),
-		"-i", wavPath, "-t", strconv.Itoa(lengthSec), "-f", "wav", "-")
-	fpcalcCmd := exec.Command("fpcalc", "-raw", "-length", strconv.Itoa(lengthSec), "-")
-
-	pr, pw := io.Pipe()
-	ffmpegCmd.Stdout = pw
-	fpcalcCmd.Stdin = pr
-
-	var out bytes.Buffer
-	fpcalcCmd.Stdout = &out
-
-	if err := ffmpegCmd.Start(); err != nil {
-		return nil, fmt.Errorf("ffmpeg: start: %w", err)
-	}
-	if err := fpcalcCmd.Start(); err != nil {
-		_ = ffmpegCmd.Process.Kill()
-		return nil, fmt.Errorf("fpcalc: start: %w", err)
-	}
-
-	ffmpegErr := ffmpegCmd.Wait()
-	pw.Close()
-	fpcalcErr := fpcalcCmd.Wait()
-
-	if ffmpegErr != nil {
-		return nil, fmt.Errorf("ffmpeg: %w", ffmpegErr)
-	}
-	if fpcalcErr != nil {
-		return nil, fmt.Errorf("fpcalc: %w", fpcalcErr)
-	}
-	return parseFpcalcOutput(out.Bytes())
-}
-
-// parseFpcalcOutput extracts the fingerprint from fpcalc -raw output.
-func parseFpcalcOutput(out []byte) (Fingerprint, error) {
-	// Output format:
-	//   DURATION=23
-	//   FINGERPRINT=1234567,9876543,...
-	for _, line := range strings.Split(string(out), "\n") {
-		if strings.HasPrefix(line, "FINGERPRINT=") {
-			return ParseFingerprint(strings.TrimPrefix(line, "FINGERPRINT="))
-		}
-	}
-	return nil, fmt.Errorf("fpcalc: FINGERPRINT line not found in output")
-}
-
-// newFingerprinter returns a fpcalcFingerprinter if both fpcalc and ffmpeg
-// are found in PATH, or nil otherwise. A nil Fingerprinter disables local
-// fingerprint caching without preventing the rest of the recognizer from
-// functioning. Both tools are required because multi-window fingerprinting
-// uses ffmpeg to trim audio segments when offsetSec > 0.
 func newFingerprinter() Fingerprinter {
-	if _, err := exec.LookPath("fpcalc"); err != nil {
-		return nil
-	}
-	if _, err := exec.LookPath("ffmpeg"); err != nil {
-		return nil
-	}
-	return fpcalcFingerprinter{}
-}
-
-// GenerateFingerprints generates up to maxWindows fingerprints from wavPath,
-// sampling the audio at stride-second intervals starting at offset 0. Windows
-// that would start at or beyond captureSec are skipped. Individual window
-// failures are logged and skipped; partial results are returned to the caller.
-func GenerateFingerprints(fp Fingerprinter, wavPath string, maxWindows, strideSec, lengthSec, captureSec int) []Fingerprint {
-	if fp == nil || maxWindows <= 0 || strideSec <= 0 || lengthSec <= 0 {
-		return nil
-	}
-	var results []Fingerprint
-	for i := 0; i < maxWindows; i++ {
-		offset := i * strideSec
-		if offset >= captureSec {
-			break
-		}
-		f, err := fp.Generate(wavPath, offset, lengthSec)
-		if err != nil {
-			log.Printf("fingerprint: window offset=%ds: %v", offset, err)
-			continue
-		}
-		results = append(results, f)
-	}
-	return results
+	return internalrecognition.NewFingerprinter()
 }
